# PlaceCar_2024_Angular/public/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PlaceCar API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Reservations/Client/NewReservation": {
      "post": {
        "tags": [
          "Reservation"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddResDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddResDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddResDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Reservations/Client/{clientId}/NonCloture": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadResClientDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadResClientDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadResClientDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Reservations/Parking/{parkingId}/NonCloture": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "name": "parkingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadReservationParkingDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadReservationParkingDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadReservationParkingDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Reservations/Client/{clientId}/Cloture": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadResClientDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadResClientDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadResClientDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Reservations/Client/{clientId}/Reservation/{reservationId}/Cloture": {
      "put": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Reservations/Client/{clientId}": {
      "put": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClientResDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClientResDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClientResDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Reservations/Client/{clientId}/Reservation/{reservationId}/Suprimer": {
      "delete": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Reservations/Statistics-Depart-Arrive/Parking/{parkingid}": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "name": "parkingid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StatReservationDebFinBO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatReservationDebFinBO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatReservationDebFinBO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reservations/Count-Nombe-Reservations/Parking/{parkingid}": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "name": "parkingid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reservations/Count-Nombe-Reservations-Total-Du-Mois/Parking/{parkingid}": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "name": "parkingid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddResDTO": {
        "type": "object",
        "properties": {
          "reS_DateDebut": {
            "type": "string",
            "format": "date-time"
          },
          "reS_DateFin": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "placeId": {
            "type": "integer",
            "format": "int32"
          },
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "formPrixId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ReadResClientDTO": {
        "type": "object",
        "properties": {
          "reS_Id": {
            "type": "integer",
            "format": "int32"
          },
          "reS_DateReservation": {
            "type": "string",
            "format": "date-time"
          },
          "reS_DateDebut": {
            "type": "string",
            "format": "date-time"
          },
          "reS_DateFin": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "plA_Etage": {
            "type": "integer",
            "format": "int32"
          },
          "plA_NumeroPlace": {
            "type": "integer",
            "format": "int32"
          },
          "parK_Id": {
            "type": "integer",
            "format": "int32"
          },
          "parK_Nom": {
            "type": "string",
            "nullable": true
          },
          "adrS_Numero": {
            "type": "integer",
            "format": "int32"
          },
          "adrS_NomRue": {
            "type": "string",
            "nullable": true
          },
          "adrS_Ville": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReadReservationParkingDTO": {
        "type": "object",
        "properties": {
          "reS_Id": {
            "type": "integer",
            "format": "int32"
          },
          "reS_DateDebut": {
            "type": "string",
            "format": "date-time"
          },
          "reS_DateFin": {
            "type": "string",
            "format": "date-time"
          },
          "plA_Id": {
            "type": "integer",
            "format": "int32"
          },
          "perS_Nom": {
            "type": "string",
            "nullable": true
          },
          "perS_Prenom": {
            "type": "string",
            "nullable": true
          },
          "perS_Id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "StatReservationDebFinBO": {
        "type": "object",
        "properties": {
          "reservationsCommencent": {
            "type": "integer",
            "format": "int32"
          },
          "reservationsTerminent": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateClientResDTO": {
        "type": "object",
        "properties": {
          "reS_Id": {
            "type": "integer",
            "format": "int32"
          },
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "reS_DateFin": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "placeId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "formPrixId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Encoder votre toke; Exemple : Bearer {token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}