# PlaceCar_2024_Angular/src/app/main/api/generated/model/readResClientDTO.ts
/**
 * PlaceCar API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ReadResClientDTO { 
    reS_Id?: number;
    reS_DateReservation?: string;
    reS_DateDebut?: string;
    reS_DateFin?: string | null;
    plA_Etage?: number;
    plA_NumeroPlace?: number;
    parK_Id?: number;
    parK_Nom?: string | null;
    adrS_Numero?: number;
    adrS_NomRue?: string | null;
    adrS_Ville?: string | null;
}