# PlaceCar_2024_Angular/src/app/main/api/generated/model/readReservationParkingDTO.ts
/**
 * PlaceCar API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ReadReservationParkingDTO { 
    reS_Id?: number;
    reS_DateDebut?: string;
    reS_DateFin?: string;
    plA_Id?: number;
    perS_Nom?: string | null;
    perS_Prenom?: string | null;
    perS_Id?: number;
}